+++ fibjs/include/MemoryStream.h
@@ -29,7 +29,7 @@ public:
 		CloneStream(std::string buffer, date_t tm) :
 			m_buffer(buffer), m_time(tm), m_pos(0)
 		{
-			extMemory(m_buffer.length());
+			extMemory((int)m_buffer.length());
 		}
 
 	public:
